+++ src/test/scala/com/sparkutils/qualityTests/BaseFunctionalityTest.scala
@@ -290,7 +290,7 @@ class BaseFunctionalityTest extends FunSuite with RowTools with TestUtils {
 
   // 2.4 doesn't support forceInterpreted so we can't test that it _doesn't_ compile, databricks is cluster based so we'll not be able to capture it without dumping to files
   @Test
-  def testPrintCode(): Unit = not_Databricks{ not2_4 {
+  def testPrintCode(): Unit = not_Cluster{ not2_4 {
     // using eval we shouldn't get output
     forceInterpreted {
       doTestPrint(null, "my message is", null, null, "printCode")
@@ -344,7 +344,7 @@ class BaseFunctionalityTest extends FunSuite with RowTools with TestUtils {
 
     // should fail
     try {
-      (df union df2 distinct).show
+      (df union df2 distinct).head//show
       fail("Is assumed to fail as spark doesn't order maps")
     } catch {
       case t: Throwable =>
@@ -358,7 +358,7 @@ class BaseFunctionalityTest extends FunSuite with RowTools with TestUtils {
     //comparable.show
     val unioned = comparable union comparable2 distinct
 
-    unioned.show
+    unioned.head
     assert(unioned.count == df.count)
   }
 
@@ -372,7 +372,7 @@ class BaseFunctionalityTest extends FunSuite with RowTools with TestUtils {
 
       // should fail
       try {
-        (df union df2 distinct).show
+        (df union df2 distinct).head
         fail("Is assumed to fail as spark doesn't order maps")
       } catch {
         case t: Throwable =>
@@ -386,13 +386,13 @@ class BaseFunctionalityTest extends FunSuite with RowTools with TestUtils {
       //comparable.show
       val unioned = comparable union comparable2 distinct
 
-      unioned.show
+      unioned.head
       if (thereCanBeOnlyOne)
         assert(unioned.count == 1)
       else
         assert(unioned.count == df.count)
 
-      unioned.sort("v").show
+      debug(unioned.sort("v").show)
     }
 
     // arrays
@@ -433,7 +433,7 @@ class BaseFunctionalityTest extends FunSuite with RowTools with TestUtils {
 
     // should fail
     try {
-      (ds1 union ds2 distinct).show
+      (ds1 union ds2 distinct).head
       fail("Is assumed to fail as spark doesn't order maps")
     } catch {
       case t: Throwable =>
@@ -447,7 +447,7 @@ class BaseFunctionalityTest extends FunSuite with RowTools with TestUtils {
     //comparable.show
     val unioned = comparable union comparable2 distinct
 
-    unioned.show
+    unioned.head
     assert(unioned.count == 1) // because all 4 are identical
   }
 
@@ -460,11 +460,11 @@ class BaseFunctionalityTest extends FunSuite with RowTools with TestUtils {
 
     import sparkSession.implicits._
     val ds = maps.reverse.map(m => MapArray(Seq(m.toMap))).toDS()
-    ds.show
+    ds.head
 
     // should fail
     try {
-      ds.sort("seq").show
+      ds.sort("seq").head
       fail("Is assumed to fail as spark doesn't order maps")
     } catch {
       case t: Throwable =>
@@ -475,7 +475,7 @@ class BaseFunctionalityTest extends FunSuite with RowTools with TestUtils {
     val comparable = ds.toDF.selectExpr("comparableMaps(seq) seq")
 
     val sorted = comparable.sort("seq").selectExpr("reverseComparableMaps(seq) seq").as[MapArray]
-    sorted.show
+    sorted.head
     sorted.collect().zipWithIndex.foreach{
       case (map,index) =>
         assert(map.seq(0) == maps(index).toMap) // because all 4 are identical
@@ -496,11 +496,11 @@ class BaseFunctionalityTest extends FunSuite with RowTools with TestUtils {
 
     import sparkSession.implicits._
     val ds = maps.reverse.map(m => NestedMapStruct(NestedStruct(m.head._1, Map( m.head._1 -> MapArray(Seq(m.toMap)))))).toDS()
-    ds.show
+    ds.head
 
     // should fail
     try {
-      ds.sort("nested").show
+      ds.sort("nested").head
       fail("Is assumed to fail as spark doesn't order maps")
     } catch {
       case t: Throwable =>
@@ -511,7 +511,7 @@ class BaseFunctionalityTest extends FunSuite with RowTools with TestUtils {
     val comparable = ds.toDF.selectExpr("comparableMaps(nested) seq")
 
     val sorted = comparable.sort("seq").selectExpr("reverseComparableMaps(seq) nested").as[NestedMapStruct]
-    sorted.show
+    sorted.head
     sorted.collect().zipWithIndex.foreach {
       case (struct, index) =>
         assert(struct.nested.nested.head._2.seq(0) == maps(index).toMap) // because all 4 are identical
@@ -740,7 +740,7 @@ class BaseFunctionalityTest extends FunSuite with RowTools with TestUtils {
   @Test
   def checkMinimumLengthWorks(): Unit =
     try {
-      sparkSession.range(1).selectExpr("hash_with()").show
+      sparkSession.range(1).selectExpr("hash_with()").head
       fail("should have thrown")
     } catch {
       case t: Throwable if t.getMessage.contains("A minimum of 2 parameters is required") =>